add test metric reader utilities
